+++ retail_sales_pipeline/models/gold/gold_sales_by_category.sql
@@ -7,8 +7,8 @@ select
     sale_month,
     count(distinct sale_id)             as total_transactions,
     sum(quantity)                       as total_units_sold,
-    round(sum(total_sale_amount), 2)    as total_revenue,
-    round(avg(unit_price), 2)           as avg_unit_price
+round(sum(total_sale_amount)::numeric, 2)    as total_revenue,
+round(avg(unit_price)::numeric, 2)           as avg_unit_price
 from {{ ref('silver_sales') }}
 where record_status = 'VALID'
 group by product_category, sale_year, sale_month
